google_cloud/src/requirements.txt
functions-framework
google-auth
google-cloud-logging
requests
